+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added `TextureBuilder::empty` constructor
 - Added `sys::Gl::set_uniform_1f` method
 - Added `sys::CullFace::Front` variant
 - Added `sys::FrontFace::CounterClockWise` variant